requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
asyncpg==0.18.3
attrs==19.1.0
blinker==1.4
Brlapi==0.6.6
Cerberus==1.3.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cli-ui==0.9.1
Click==7.0
cloudpickle==1.2.2
colorama==0.3.7
colored-traceback==0.3.0
command-not-found==0.3
contextlib2==0.5.5
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dask==2.6.0
decorator==4.4.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Eve==0.9.2
Events==0.3
Flask==1.1.1
github3.py==1.3.0
httplib2==0.9.2
idna==2.6
imageio==2.6.1
importlib-metadata==0.19
isort==4.3.20
itsdangerous==0.24
Jinja2==2.10.1
jwcrypto==0.6.0
keras-datasets==0.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
netifaces==0.10.4
networkx==2.4
numpy==1.15.0
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.0.25
path.py==12.0.1
pbr==5.4.0
pexpect==4.2.1
Pillow==5.1.0
ply==3.11
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pyinotify==0.9.6
pylint==2.3.1
pymacaroons==0.13.0
pymongo==3.9.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
python-gitlab==1.10.0
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
schema==0.7.0
scikit-image==0.16.2
scipy==1.3.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
SQLAlchemy==1.1.11
stevedore==1.30.1
system-service==0.3
systemd-python==234
tabulate==0.8.3
toolz==0.10.0
tsrc==0.9.0
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wadllib==1.3.2
Werkzeug==0.15.4
wrapt==1.11.2
xattr==0.9.2
xkit==0.0.0
yapf==0.28.0
zipp==0.6.0
zope.interface==4.3.2
